use default reward addr in cli
